Crate mem_map

Library for copying certain memory regions and storing their pointers inside a hash map collection.

MemMap offers also a simple way of IPC.
All processes implementing this library can access each other’s maps without any blocking or security.

Usage

This lib can be used as Rust crate or system library

[dependencies.mem-map]
path = "https://github.com/VulcoLibs/mem-map"

To build a system library, simply go to mem-map’s Cargo.toml file and follow the instructions.

Inner Functions

Inner Functions are functions that work at own process’ map.

init

insert_cp

insert_mv

get

get_from_reg

remove

Extern Functions

Extern Functions are functions that work at specified process’ map.

get_extern

insert_extern

id_exists

Modules

errors

exports

mem_map

Constants

REG_PATH

Registry path to the MemMap keys.
Full path: HKEY_CURRENT_USER\SOFTWARE\VulcoLibs\MemMap

Functions

get⚠

get - Gets a pointer to saved data

get_extern⚠

get_extern - Saves a data from another process into buffer.

get_from_reg⚠

Deprecated
get_from_reg - Gets a pointer to saved data by registry entry.

id_exists⚠

id_exists - Checks if certain map key (ID) exists.

init⚠

init - Initializes the mem-map library.

insert_cp⚠

insert_cp - Copies a certain memory region into an own map

insert_extern⚠

insert_extern - Copies a certain memory region into a map of another process.

insert_mv⚠

insert_mv - Copies a certain memory region into an own map

remove⚠

remove - Removes the saved data