Module mem_map::exports

Functions

get⚠

get - Gets a pointer to saved data

get_extern⚠

get_extern - Saves a data from another process into buffer.

get_from_reg⚠

Deprecated
get_from_reg - Gets a pointer to saved data by registry entry.

id_exists⚠

id_exists - Checks if certain map key (ID) exists.

init⚠

init - Initializes the mem-map library.

insert_cp⚠

insert_cp - Copies a certain memory region into an own map

insert_extern⚠

insert_extern - Copies a certain memory region into a map of another process.

insert_mv⚠

insert_mv - Copies a certain memory region into an own map

remove⚠

remove - Removes the saved data